Switch to Gradle + reorganize files
