+++ deadchest-core/src/main/java/me/crylonz/PermissionUtils.java
