+++ deadchest-core/src/main/java/me/crylonz/Utils.java
