+++ deadchest-core/src/main/java/me/crylonz/WorldGuardSoftDependenciesChecker.java
