+++ deadchest-core/src/main/java/me/crylonz/commands/DCCommandRegistration.java
