+++ deadchest-core/src/main/java/me/crylonz/commands/DCCommandRegistrationService.java
