+++ deadchest-core/src/main/java/me/crylonz/utils/ConfigKey.java
