+++ deadchest-core/src/main/java/me/crylonz/utils/DeadChestConfig.java
