+++ deadchest-core/src/main/java/me/crylonz/utils/DeadChestUpdater.java
