+++ deadchest-core/src/main/resources/plugin.yml
@@ -1,0 +1,42 @@
+name: DeadChest
+main: me.crylonz.DeadChest
+softdepend: [Multiverse-Core, WorldGuard]
+author: Crylonz
+version: "4.16.0"
+description: Keep your inventory in a chest when you die
+api-version: "1.13"
+permissions:
+  deadchest.generate:
+    description: Allows a player to keep inventory in a chest when he dies
+    default: op
+  deadchest.get:
+    description: Allows a player to get his inventory by picking his deadchest
+    default: op
+  deadchest.chestPass:
+    description: Player can open all deadChest
+    default: op
+  deadchest.infinityChest:
+    description: Player dead chest never disappaear
+    default: false
+  deadchest.giveBack:
+    description: Allow a player to give back a deadchest to a player
+    default: op
+  deadchest.admin:
+    description: Player dead chest never disappaear
+    default: op
+  deadchest.remove.own:
+    description: Allow a player to remove his deadchest
+    default: op
+  deadchest.remove.other:
+    description: Allow a player to remove others players deadchest
+    default: op
+  deadchest.list.own:
+    description: Allow a player to list his deadchests
+    default: op
+  deadchest.list.other:
+    description: Allow a player to list other deadchests
+    default: op
+commands:
+  dc:
+    usage: /dc reload | repair | remove [PlayerName] | removeinfinite | removeall | list [PlayerName] | list all
+    description: List of commands
