--- src/main/resources/config.yml
@@ -1,102 +1,0 @@
-################################################################################################################################
-# Welcome to the DeadChest configuration file!                                                                                 #
-#                                                                                                                              #
-# You can find more information about the plugin and this file on the plugin page:                                             #
-#   > https://dev.bukkit.org/projects/dead-chest                                                                               #
-# If you need any support, feel free to ask for it on our Discord server:                                                      #
-#   > https://discordapp.com/invite/jCsvJxS                                                                                    #
-#                                                                                                                              #
-# If you make any changes in this file, please run the command '/dc reload' to apply the changes!                              #
-################################################################################################################################
-
-auto-update: true
-
-# Should DeadChests be openable by other players or just the owner?
-#   > false - DeadChests can be opened by anyone.
-#   > true  - DeadChests can only be opened by the owner.
-OnlyOwnerCanOpenDeadChest: true
-
-# The amount of time a DeadChest will be active for (in seconds)...
-#   You can disable this limit by setting the value to 0.
-DeadChestDuration: 300
-
-# Should DeadChests be protected from explosions and other players?
-#   > false - DeadChests will not be protected and can be broken.
-#   > true  - DeadChests will be protected.
-IndestructibleChest: true
-
-# The maximum number of DeadChests a player can have active at a time...
-#   You can disable this limit by setting the value to 0.
-maxDeadChestPerPlayer: 15
-
-# Should we notify the console when and where a DeadChest is created?
-logDeadChestOnConsole: false
-
-# Should players require permission ('deadchest.generate') to generate a DeadChest?
-#   > false - Players will NOT require permission to generate a DeadChest.
-#   > true  - Players WILL require permission to generate a DeadChest.
-RequirePermissionToGenerate: false
-
-# Should players require permission ('deadchest.get') to retrieve their DeadChest?
-#   > false - Players will NOT require permission to retrieve their DeadChest.
-#   > true  - Players WILL require permission to retrieve their DeadChest.
-RequirePermissionToGetChest: false
-
-# Should players require permission ('deadchest.list.own') to view their DeadChests (using '/dc list')?
-#   > false - Players will NOT require permission to view their DeadChests.
-#   > true  - Players WILL require permission to view their DeadChests.
-RequirePermissionToListOwn: false
-
-# Should we remove all DeadChests when the server starts?
-AutoCleanupOnStart: false
-
-# Should players in creative mode create a DeadChest when they die?
-GenerateDeadChestInCreative: true
-
-# Should we tell the player where their DeadChest is located when they die?
-DisplayDeadChestPositionOnDeath: true
-
-# How should we return the items to the player when they claim their DeadChest?
-#   > 1 - Items are added directly to the players inventory until it's full, excess gets dropped.
-#   > 2 - All items are dropped on the ground.
-DropMode: 1
-
-# What block should we use for the DeadChest?
-#   > 1 - CHEST
-#   > 2 - PLAYER HEAD
-#   > 3 - BARREL
-#   > 4 - SHULKER BOX
-#   > 5 - ENDER CHEST
-DropBlock: 1
-
-# Store Player XP on Deadchest
-StoreXP: false
-
-# Should we drop DeadChest contents when the timer expires?
-#   > false - DeadChest contents are removed when the timer expires.
-#   > true  - DeadChest contents are dropped when the timer expires.
-ItemsDroppedAfterTimeOut: false
-
-# Should generate a DeadChest on lava?
-GenerateOnLava: true
-# Should generate a DeadChest on water?
-GenerateOnWater: true
-# Should generate a DeadChest on rails?
-GenerateOnRails: true
-# Should generate a DeadChest in minecarts?
-GenerateInMinecart: true
-
-# Should we check for WorldGuard regions when a player dies?
-#   > false - WorldGuard regions will NOT be checked for.
-#   > true  - WorldGuard regions WILL be checked for.
-EnableWorldGuardDetection: false
-
-# A list of worlds that DeadChests will NOT be generated in...
-ExcludedWorld:
-  - 'excluded_world_example'
-  - 'another_excluded_world_example'
-
-# A list of items that will be excluded from DeadChests...
-ExcludedItems:
-  - 'EXCLUDED_ITEM_EXAMPLE'
-  - 'HEROBRINE'
